Align with develop structure: rename Extension to CommEx and remove unused plugins
