+++ BveExCsTemplate.sln
@@ -3,11 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33516.290
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VehiclePlugin", "VehiclePlugin\VehiclePlugin.csproj", "{B4C723A7-FD6A-4931-B3BF-A6635F44899B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MapPlugin", "MapPlugin\MapPlugin.csproj", "{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extension", "Extension\Extension.csproj", "{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommEx", "CommEx\CommEx.csproj", "{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extension.Tests", "Extension.Tests\Extension.Tests.csproj", "{326BBA13-B6C3-4779-80BF-AB31E801A5A5}"
 EndProject
@@ -17,14 +13,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B4C723A7-FD6A-4931-B3BF-A6635F44899B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B4C723A7-FD6A-4931-B3BF-A6635F44899B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B4C723A7-FD6A-4931-B3BF-A6635F44899B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B4C723A7-FD6A-4931-B3BF-A6635F44899B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}.Release|Any CPU.ActiveCfg = Release|Any CPU
