+++ CommEx/CommEx.csproj
@@ -3,7 +3,7 @@
 
   <PropertyGroup>
     <TargetFramework>net48</TargetFramework>
-    <RootNamespace>BveExCsTemplate.VehiclePlugin</RootNamespace>
+    <RootNamespace>CommEx</RootNamespace>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
     <Deterministic>false</Deterministic>
     <RestorePackagesWithLockFile>true</RestorePackagesWithLockFile>
@@ -11,6 +11,7 @@
 
   <ItemGroup>
     <PackageReference Include="BveEx.CoreExtensions" Version="2.0.8" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 
 </Project>
