+++ CommEx/Infrastructure/UdpRelayService.cs
@@ -2,9 +2,9 @@ using System;
 using System.Text;
 using Newtonsoft.Json;
 
-using BveExCsTemplate.Extension.Model;
+using CommEx.Model;
 
-namespace BveExCsTemplate.Extension.Infrastructure
+namespace CommEx.Infrastructure
 {
     internal interface IUdpSender : IDisposable
     {
