+++ Extension.Tests/Extension.Tests.csproj
@@ -14,14 +14,14 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\Extension\Model\SimulationFrame.cs" Link="Model\SimulationFrame.cs" />
-    <Compile Include="..\Extension\Model\BveExModelStore.cs" Link="Model\BveExModelStore.cs" />
-    <Compile Include="..\Extension\Infrastructure\NtpServerService.cs" Link="Infrastructure\NtpServerService.cs" />
-    <Compile Include="..\Extension\Infrastructure\UdpRelayService.cs" Link="Infrastructure\UdpRelayService.cs" />
-    <Compile Include="..\Extension\Infrastructure\ApiBridgeService.cs" Link="Infrastructure\ApiBridgeService.cs" />
-    <Compile Include="..\Extension\Infrastructure\CommunicationWorker.cs" Link="Infrastructure\CommunicationWorker.cs" />
-    <Compile Include="..\Extension\Infrastructure\IFramePublisher.cs" Link="Infrastructure\IFramePublisher.cs" />
-    <Compile Include="..\Extension\Infrastructure\SerialRelayService.cs" Link="Infrastructure\SerialRelayService.cs" />
-    <Compile Include="..\Extension\Infrastructure\BidsProtocol.cs" Link="Infrastructure\BidsProtocol.cs" />
+    <Compile Include="..\CommEx\Model\SimulationFrame.cs" Link="Model\SimulationFrame.cs" />
+    <Compile Include="..\CommEx\Model\BveExModelStore.cs" Link="Model\BveExModelStore.cs" />
+    <Compile Include="..\CommEx\Infrastructure\NtpServerService.cs" Link="Infrastructure\NtpServerService.cs" />
+    <Compile Include="..\CommEx\Infrastructure\UdpRelayService.cs" Link="Infrastructure\UdpRelayService.cs" />
+    <Compile Include="..\CommEx\Infrastructure\ApiBridgeService.cs" Link="Infrastructure\ApiBridgeService.cs" />
+    <Compile Include="..\CommEx\Infrastructure\CommunicationWorker.cs" Link="Infrastructure\CommunicationWorker.cs" />
+    <Compile Include="..\CommEx\Infrastructure\IFramePublisher.cs" Link="Infrastructure\IFramePublisher.cs" />
+    <Compile Include="..\CommEx\Infrastructure\SerialRelayService.cs" Link="Infrastructure\SerialRelayService.cs" />
+    <Compile Include="..\CommEx\Infrastructure\BidsProtocol.cs" Link="Infrastructure\BidsProtocol.cs" />
   </ItemGroup>
 </Project>
