+++ Extension.Tests/ModelAndTransportTests.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading;
 
-using BveExCsTemplate.Extension.Infrastructure;
-using BveExCsTemplate.Extension.Model;
+using CommEx.Infrastructure;
+using CommEx.Model;
 
 using Newtonsoft.Json.Linq;
