--- MapPlugin/MapPlugin.csproj
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <RootNamespace>BveExCsTemplate.MapPlugin</RootNamespace>
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-    <Deterministic>false</Deterministic>
-    <RestorePackagesWithLockFile>true</RestorePackagesWithLockFile>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="BveEx.CoreExtensions" Version="2.0.8" />
-  </ItemGroup>
-
-</Project>
